+++ test/dummy/test/integration/recording_studio_v3_host_test.rb
@@ -8,7 +8,7 @@ class RecordingStudioV3HostTest < ActiveSupport::TestCase
     assert_equal :application_layout, RecordingStudioRootSwitchable.configuration.layout
     assert_includes ApplicationController.ancestors, RecordingStudio::RootSwitchable::ControllerSupport
     refute_includes ApplicationController.ancestors, RecordingStudio::UsesDefaultLayout
-    assert_includes RecordingStudioAI::Admin::ApplicationController.ancestors, RecordingStudio::UsesDefaultLayout
+    assert_includes RecordingStudioAI::RetainedResponsesController.ancestors, RecordingStudio::UsesDefaultLayout
   end
 
   test "dummy app validates v3 recordable declarations" do